Refactor: Moved Util Functions to Twitter App
